utils/requirements.txt
google-cloud-storage  # For interacting with GCS
# google-cloud-bigquery  # For BigQuery operations
# google-cloud-dataproc  # For Dataproc cluster and job management
# google-cloud-pubsub  # Optional: If using Pub/Sub triggers
# google-cloud-dataflow  # If you are using Dataflow with Python SDK